requirements.txt
nltk==3.8.1
pandas==2.1.4
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.3
emoji==2.10.1
tqdm==4.66.2
contractions==0.1.73
joblib==1.3.2
scikit-learn
spacy
imbalanced-learn
pyyaml